+++ docs/changelog.rst
@@ -2,6 +2,8 @@
 Changelog
 =========
 
+* :feature:`10980` Users will now be able to report unsupported events through the history event page.
+* :feature:`6642` Users will now be able to provide feedback directly though the applications help & support sidebar.
 * :bug:`11099` rotki will now allow to create events manually with location Avalanche.
 * :bug:`11002` Liquity V2 proxy detection will now work correctly and proxy deployment transactions will be decoded properly.
 * :feature:`-` StakeDAO V2 Curve strategy transactions will be correctly decoded with vault balances properly detected.